restoredoc/requirements.txt
# Flask Server Requirements for RestoreDoc
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional but recommended
openai>=1.6.1